TextParser/requirements.txt
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.5.0
pytesseract>=0.3.8
imutils>=0.5.4
tqdm>=4.65.0 